+++ Assets/Plugins/GameJolt/Scripts/API/DataStore.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 using GameJolt.API.Core;
 
 namespace GameJolt.API {
@@ -21,13 +23,17 @@ namespace GameJolt.API {
 	public static class DataStore {
 		/// <summary>
 		/// GameJolt refuses requests larger than this.
-		/// In order for an operation to succeed the key length + value length must not be larger than this.
+		/// In order for an operation to succeed the encoded key length + encoded value length must not be larger than this.
 		/// </summary>
-		public const int SoftLimit = 1024 * 1024 - 10;
+		public const int SoftLimit = 1024 * 1024 - 11;
 		/// <summary>
 		/// GameJolt allows only about 16MB per key-value pair.
 		/// </summary>
 		public const int HardLimit = 16777000; //~16MB
+		/// <summary>
+		/// Maximum size for dataset key.
+		/// </summary>
+		public const int KeySizeLimit = SoftLimit - 1024;
 
 		/// <summary>
 		/// Save the key/value pair in the DataStore.
@@ -40,7 +46,7 @@ namespace GameJolt.API {
 		/// <param name="global">A boolean indicating whether the key is global (<c>true</c>) or private to the user (<c>false</c>).</param>
 		/// <param name="callback">A callback function accepting a single parameter, a boolean indicating success.</param>
 		public static void Set(string key, string value, bool global, Action<bool> callback = null) {
-			if(key.Length + value.Length > SoftLimit) {
+			if(GetEncodedSize(key) + GetEncodedSize(value) > SoftLimit) {
 				LogHelper.Error("Failed to upload data, because it was to large.");
 				if(callback != null) callback(false);
 				return;
@@ -74,29 +80,31 @@ namespace GameJolt.API {
 			Action<int> progress = null, int maxSegmentSize = SoftLimit) {
 			if(callback == null) throw new ArgumentNullException();
 			if(maxSegmentSize < 10 || maxSegmentSize > SoftLimit) throw new ArgumentOutOfRangeException();
-			if(key.Length + value.Length > HardLimit) {
-				LogHelper.Error("Datset is larger than 16MB!");
-				callback(false);
-				return;
-			}
-			if(key.Length >= SoftLimit - 1024) {
+
+			int encodedKeySize = GetEncodedSize(key);
+			if(encodedKeySize >= KeySizeLimit) {
 				LogHelper.Error("Failed to upload data, because the key is too long.");
 				callback(false);
 				return;
 			}
+			maxSegmentSize = Math.Min(maxSegmentSize, SoftLimit - encodedKeySize);
+			var segments = new Queue<string>();
+			int encodedSize = Segmentate(Encoding.ASCII.GetBytes(value), maxSegmentSize, segments);
+			if(encodedKeySize + encodedSize > HardLimit) {
+				LogHelper.Error("Dataset is larger than 16MB!");
+				callback(false);
+				return;
+			}
 			const int keyLimit = SoftLimit * 3 / 4;
-			if(key.Length > keyLimit) {
+			if(encodedKeySize > keyLimit) {
 				LogHelper.Warning("Key is very long, only {0} bytes left for the data. " +
 				                  "Therefore many segments may be needed to upload the data." +
 				                  "Consider using smaller keys.", SoftLimit - key.Length);
 			}
 
 			// set first segment
-			maxSegmentSize = Math.Min(maxSegmentSize, SoftLimit - key.Length);
-			maxSegmentSize = Math.Min(maxSegmentSize, value.Length);
-			var segmentSize = maxSegmentSize;
 			var dataSend = 0;
-			var segment = value.Substring(0, segmentSize);
+			var segment = segments.Dequeue();
 			var payload = new Dictionary<string, string> {{"key", key}, {"data", segment}};
 
 			Action<string> completedAction = null;
@@ -104,13 +112,12 @@ namespace GameJolt.API {
 				if(response == null) { // request failed
 					callback(false);
 				} else { // request succeeded
-					dataSend += segmentSize;
+					dataSend += segment.Length;
 					if(progress != null) progress(dataSend);
 					if(dataSend >= value.Length) { // data uploaded completely
 						callback(true);
 					} else { // append next segment
-						segmentSize = Math.Min(maxSegmentSize, value.Length - dataSend);
-						segment = value.Substring(dataSend, segmentSize);
+						segment = segments.Dequeue();
 						Update(key, segment, DataStoreOperation.Append, global, completedAction);
 					}
 				}
@@ -130,8 +137,8 @@ namespace GameJolt.API {
 		/// <param name="callback">A callback function accepting a single parameter, a the updated key value.</param>
 		public static void Update(string key, string value, DataStoreOperation operation, bool global,
 			Action<string> callback = null) {
-			var parameters = new Dictionary<string, string> {{"key", key}, {"operation", operation.ToString().ToLower()}};
-			var payload = new Dictionary<string, string> {{"value", value}};
+			var parameters = new Dictionary<string, string> {{"operation", operation.ToString().ToLower()}};
+			var payload = new Dictionary<string, string> {{"key", key}, {"value", value}};
 
 			Request.Post(Constants.ApiDatastoreUpdate, parameters, payload, response => {
 				if(callback != null) {
@@ -147,9 +154,9 @@ namespace GameJolt.API {
 		/// <param name="global">A boolean indicating whether the key is global (<c>true</c>) or private to the user (<c>false</c>).</param>
 		/// <param name="callback">A callback function accepting a single parameter, the key value.</param>
 		public static void Get(string key, bool global, Action<string> callback) {
-			var parameters = new Dictionary<string, string> {{"key", key}};
+			var payload = new Dictionary<string, string> {{"key", key}};
 
-			Request.Get(Constants.ApiDatastoreFetch, parameters, response => {
+			Request.Post(Constants.ApiDatastoreFetch, null, payload, response => {
 				var value = response.Success ? response.Dump : null;
 				if(callback != null) {
 					callback(value);
@@ -164,9 +171,9 @@ namespace GameJolt.API {
 		/// <param name="global">A boolean indicating whether the key is global (<c>true</c>) or private to the user (<c>false</c>).</param>
 		/// <param name="callback">A callback function accepting a single parameter, a boolean indicating success.</param>
 		public static void Delete(string key, bool global, Action<bool> callback = null) {
-			var parameters = new Dictionary<string, string> {{"key", key}};
+			var payload = new Dictionary<string, string> {{"key", key}};
 
-			Request.Get(Constants.ApiDatastoreRemove, parameters, response => {
+			Request.Post(Constants.ApiDatastoreRemove, null, payload, response => {
 				if(callback != null) {
 					callback(response.Success);
 				}
@@ -210,5 +217,47 @@ namespace GameJolt.API {
 				}
 			}, !global);
 		}
+
+		#region Utils
+		private static readonly byte[] EncodedSize = new byte[256];
+
+		static DataStore() {
+			for(char c = '\0'; c < 256; c++) {
+				// ASCII chars from 0-32 and 127-255 are invalid
+				// Furthermore the characters given by the string constant are also invalid
+				// Invalid characters must be percent encoded, for e.g. "!" becomes "%21"
+				EncodedSize[c] = (byte)((c <= 32 || c >= 127 || "@&;:<>=?\"'/\\!#%+$,{}|^[]`".Contains(c)) ? 3 : 1);
+			}
+		}
+
+		private static int GetEncodedSize(string text) {
+			return Encoding.ASCII.GetBytes(text).Sum(c => EncodedSize[c]);
+		}
+
+		/// <summary>
+		/// Subdivide the given message into one or multiple smaller segments.
+		/// </summary>
+		/// <param name="bytes">The data bytes to split.</param>
+		/// <param name="maxSegmentSize">The maximum size of a single segment.</param>
+		/// <param name="result">The segments will be added to this queue.</param>
+		/// <returns></returns>
+		private static int Segmentate(byte[] bytes, int maxSegmentSize, Queue<string> result) {
+			int start = 0; // start of the current segment
+			int capacity = 0; // url encoded size of the current segment
+			int totalCapacity = 0; // url encoded size of the whole message
+			for(int i = 0; i < bytes.Length; i++) {
+				int charSize = EncodedSize[bytes[i]]; // invalid characters must be encoded, for e.g. "!" becomes "%21"
+				capacity += charSize;
+				totalCapacity += charSize;
+				if(capacity > maxSegmentSize) {
+					result.Enqueue(Encoding.ASCII.GetString(bytes, start, i - start));
+					start = i;
+					capacity = charSize;
+				}
+			}
+			result.Enqueue(Encoding.ASCII.GetString(bytes, start, bytes.Length - start));
+			return totalCapacity;
+		}
+		#endregion
 	}
 }